+++ packages/shared/widget-plugin-sorting/src/helpers/SortStoreProvider.ts
@@ -1,4 +1,3 @@
-import { AttributeMetaData, DynamicValue } from "mendix";
 import { SortOrderStore } from "../stores/SortOrderStore";
 import { SortStoreHost } from "../stores/SortStoreHost";
 import { SortInstruction } from "../types/store";
@@ -6,10 +5,6 @@ import { SortInstruction } from "../types/store";
 interface SortStoreProviderSpec {
     host: SortStoreHost;
     initSortOrder?: SortInstruction[];
-    attributes: Array<{
-        attribute: AttributeMetaData;
-        caption?: DynamicValue<string>;
-    }>;
 }
 
 export class SortStoreProvider {
@@ -18,11 +13,7 @@ export class SortStoreProvider {
 
     constructor(spec: SortStoreProviderSpec) {
         this._host = spec.host;
-        const options = spec.attributes.map(item => ({
-            value: item.attribute.id,
-            caption: item.caption?.value ?? "empty"
-        }));
-        this.store = new SortOrderStore({ options, initSortOrder: spec.initSortOrder });
+        this.store = new SortOrderStore({ initSortOrder: spec.initSortOrder });
     }
 
     setup(): () => void {