+++ packages/pluggableWidgets/dropdown-sort-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue with Dropdown sort widget not showing attribute captions correctly.
+
 ## [1.2.2] - 2025-03-31
 
 ### Fixed